More spatial tests and taking into account review comments.
